+++ CHANGELOG.rst
@@ -16,9 +16,11 @@ v1.8
   see `#222 <https://github.com/OpenCyphal/pycyphal/issues/222>`_.
 
 - **v1.8.3:**
+
   - ``DiagnosticPublisher``: do not instantiate the publisher if the local node is anonymous.
   - ``publish_soon()``: Do not log error if closed.
   - Fix assertion failure during register value coercion.
+  - SocketCAN: close the media instance automatically on unrecoverable errors like ENODEV, ENXIO, EBADF, EBADFD, etc.
 
 v1.7
 ----